+++ nautilus_trader/examples/indicators/ema_python.py
@@ -39,7 +39,7 @@ class PyExponentialMovingAverage(Indicator):
     period : int
         The rolling window period for the indicator (> 0).
     price_type : PriceType
-        The specified price type for extracting values from quote ticks.
+        The specified price type for extracting values from quotes.
 
     Raises
     ------