+++ nautilus_trader/execution/trailing.pyx
@@ -68,7 +68,7 @@ cdef class TrailingStopCalculator:
                 raise RuntimeError(  # pragma: no cover (design-time error)
                     f"cannot process trailing stop, "
                     f"no LAST price for {order.instrument_id} "
-                    f"(add trade ticks or use bars)",
+                    f"(add trades or use bars)",
                 )
             if order.side == OrderSide.BUY:
                 temp_trigger_price = TrailingStopCalculator.calculate_with_last(
@@ -115,13 +115,13 @@ cdef class TrailingStopCalculator:
                 raise RuntimeError(  # pragma: no cover (design-time error)
                     f"cannot process trailing stop, "
                     f"no BID price for {order.instrument_id} "
-                    f"(add quote ticks or use bars)",
+                    f"(add quotes or use bars)",
                 )
             if ask is None:
                 raise RuntimeError(  # pragma: no cover (design-time error)
                     f"cannot process trailing stop, "
                     f"no ASK price for {order.instrument_id} "
-                    f"(add quote ticks or use bars)",
+                    f"(add quotes or use bars)",
                 )
 
             if order.side == OrderSide.BUY:
@@ -173,19 +173,19 @@ cdef class TrailingStopCalculator:
                 raise RuntimeError(  # pragma: no cover (design-time error)
                     f"cannot process trailing stop, "
                     f"no LAST price for {order.instrument_id} "
-                    f"(add trade ticks or use bars)",
+                    f"(add trades or use bars)",
                 )
             if bid is None:
                 raise RuntimeError(  # pragma: no cover (design-time error)
                     f"cannot process trailing stop, "
                     f"no BID price for {order.instrument_id} "
-                    f"(add quote ticks or use bars)",
+                    f"(add quotes or use bars)",
                 )
             if ask is None:
                 raise RuntimeError(  # pragma: no cover (design-time error)
                     f"cannot process trailing stop, "
                     f"no ASK price for {order.instrument_id} "
-                    f"(add quote ticks or use bars)",
+                    f"(add quotes or use bars)",
                 )
 
             if order.side == OrderSide.BUY: