+++ nautilus_trader/indicators/average/wma.pyx
@@ -37,7 +37,7 @@ cdef class WeightedMovingAverage(MovingAverage):
     weights : iterable
         The weights for the moving average calculation (if not ``None`` then = period).
     price_type : PriceType
-        The specified price type for extracting values from quote ticks.
+        The specified price type for extracting values from quotes.
 
     Raises
     ------