+++ nautilus_trader/persistence/wranglers_v2.py
@@ -68,7 +68,7 @@ class OrderBookDeltaDataWranglerV2(WranglerBase):
     Warnings
     --------
     This wrangler is used to build the PyO3 exposed version of `OrderBookDelta` and
-    will not work the same way as the current wranglers which build the legacy `Cython` trade ticks.
+    will not work the same way as the current wranglers which build the legacy `Cython` trades.
 
     """
 
@@ -173,7 +173,7 @@ class QuoteTickDataWranglerV2(WranglerBase):
     Warnings
     --------
     This wrangler is used to build the PyO3 exposed version of `QuoteTick` and
-    will not work the same way as the current wranglers which build the legacy `Cython` quote ticks.
+    will not work the same way as the current wranglers which build the legacy `Cython` quotes.
 
     """
 
@@ -290,7 +290,7 @@ class TradeTickDataWranglerV2(WranglerBase):
     Warnings
     --------
     This wrangler is used to build the PyO3 exposed version of `TradeTick` and
-    will not work the same way as the current wranglers which build the legacy `Cython` trade ticks.
+    will not work the same way as the current wranglers which build the legacy `Cython` trades.
 
     """
 
@@ -413,7 +413,7 @@ class BarDataWranglerV2(WranglerBase):
     Warnings
     --------
     This wrangler is used to build the PyO3 exposed version of `Bar` and
-    will not work the same way as the current wranglers which build the legacy `Cython` trade ticks.
+    will not work the same way as the current wranglers which build the legacy `Cython` trades.
 
     """